Restructure
Topmost level is now SqrtWrapper. It instantiates SQRTTop (the previous top module) and RoundingBlock. The point for this move is that the SQRTTop could be included in an FPU in an easier way by not including a hard-wired rounding block, as the FPU might have a common rounding block for all operations.